Removing unnecessary casts from main.go

+++ cmd/got/main.go
@@ -14,7 +14,7 @@ var (
 	url         string
 	version     string
 	dest        = flag.String("out", "", "Downloaded file destination.")
-	chunkSize   = flag.Int("size", 0, "Maximum chunk size in bytes.")
+	chunkSize   = flag.Uint64("size", 0, "Maximum chunk size in bytes.")
 	concurrency = flag.Uint("concurrency", 10, "Maximum chunks to download at the same time.")
 )
 
@@ -49,7 +49,7 @@ func main() {
 	d := got.Download{
 		URL:         url,
 		Dest:        *dest,
-		ChunkSize:   uint64(*chunkSize),
+		ChunkSize:   *chunkSize,
 		Interval:    100,
 		Concurrency: *concurrency,
 	}
@@ -63,10 +63,10 @@ func main() {
 
 		fmt.Printf(
 			"\r\r\bTotal: %s | Chunk: %s | Concurrency: %d | Received: %s | Time: %s | Avg: %s/s | Speed: %s/s",
-			humanize.Bytes(uint64(p.TotalSize)),
-			humanize.Bytes(uint64(d.ChunkSize)),
+			humanize.Bytes(p.TotalSize),
+			humanize.Bytes(d.ChunkSize),
 			d.Concurrency,
-			humanize.Bytes(uint64(p.Size)),
+			humanize.Bytes(p.Size),
 			p.TotalCost().Round(time.Second),
 			humanize.Bytes(p.AvgSpeed()),
 			humanize.Bytes(p.Speed()),
